SDC-16106. Fix SQS pipeline data loss on pipeline stop
In the previous implementation there was no distinction between cases
when a record reached a destination and cases when it didn't reach any
destination because the pipeline had been interrupted. But in the both
case the previous implementation committed a record (removed a message
from a queue).
With this fix we will support At least one and At most one modes,
for the At least one mode a message will not be removed from a queue
if a record doesn't reach a destination.

Change-Id: Ic392b38af141d62dcaddd60c5c47136167475b45
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36387
Reviewed-by: Xavier Baqués <xavi@streamsets.com>