SDC-16137: Health Inspector: Finalize data model
The change might look larger, but it's actually fairly small - I got
clear feedback that the model is confusing. The feature is called
"Health Inspector" and underneath there is a separate thing that is
called "Health Inspector". Thus this patch cleans the naming up a bit:

* "Health Inspector” Name of the feature
* “Health Check Category” Categories of related health checks
* “Health Report” report generated by running one or more categories of health check
* “Health Check” each individual “line” on the webpage

I've quickly run that by docs and ep to ensure that we can all agree on
the nomenclature.

Change-Id: If1d8070cca031544257013cd9d2543d6c615552c
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36035
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>