SDC-15863. Mutlitable JDBC produces duplicate records with SWITCH_TABLE
The root cause of the issue is that if one thread jumps to processing
of another table and then gets back, in between another thread could
process a batch on that first table, the second thread also could jump
from its table to the table the first thread was processing the first.
Thus when the first thread gets back to its first table it continues
where it left the table, but some records were already processed by
the second thread and the first thread is not aware of that. This way
we get duplicate records.

This issue, obviously, doesn't happen when the
"Process All Available Rows From the Table" mode is used (because
threads do not jump to other tables). In this fix we allow threads
to jump between threads, but we do not allow other threads to process
records of table that is owned by another thread.

This may have a performance impact, but there is a workaround if we
enable the table partitioning. With table partitioning several threads
may work on the same table. This is possible because threads
will exclusively own partitions, one partition will be processed by
only one thread. And if it happens that one thread processes one
partition, then jumps to another table and then jumps back, to process
the second partition of the first table that thread will use another
result set with the cursor set to the beginning of the partition.

Change-Id: I0b6a8e33da38ea47b514e6976af781758060c08d
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36091
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>