SDC-16164. Google Pub Sub Producer are failing after the SDK upgrade
After updating the SDK to the latest are some properties (Max Outstanding Message Count and Max Outstanding Request Bytes) that previously had default values in the SDK, we were able to left them empty and those values were picked up. Now, in the newest version, those defaults are gone and the value is required in our side, so we have changed the defaults, and made some changes to always set them. Apart from that we will upgrade old pipelines with value = 0, but those who set a value will remain with it.

Change-Id: Iefb5a76aa494b51f96f178a5a255a9546437a4b9
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36123
Reviewed-by: Loretta Chen <loretta@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>