SDC-16246. Cache doesn't work in Kudu Lookup Processor
Google cache compares key by equality. Unfortunately, our
implementation of KuduLookupKey doesn't override the equals method,
thus in our case the comparison will be done by identity. So if we
use another instance of the look up key Google cache will think it's
a different key and will trigger loading.

In this change set we add an implementation of the equals method.
This way even if different instances of a look up key is used,
but the both keys use the same table and the same id columns they will
be treated as the same key of a cache item.

Change-Id: Ibad3807b5f22e7568f2be64575ba240719032317
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36368
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>