SDC-9344. Kudu Lookup Processor Retry on Missing Value
We cannot use the standard loading cache because it doesn't support
the retryOnCacheMiss value set to true. But we also cannot use the
optional loading cache as is because we cannot provide default values
at init time. To get default values we need to convert them from
strings into corresponding target types which we will know only after
we get table metadata. So, to use the optional loading cache we we
give it an empty list as a default value. This empty list will never
be used since our loader always returns not empty optional value (it
may contain an empty list too, but the optional value will be not
empty). With this implementation the cache for default values will be
always on. To disable the caching for default values depending on the
retryOnCacheMiss conf value, we will invalidate manually a value that
we get from the cache if

1) it's a default value and
2) retryOnCacheMiss is ON

Change-Id: Ia717e5d5a90e49efdfd42c0b73125b4a466bd077
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36381
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hugo Prol <hugo@streamsets.com>