SDC-16295: AzureUtils.updateConfigsForConnections should take into account that authTokenEndpoint is an optional parameter
We found this in our upgrade tests - previously the authTokenEndpoint
was required only in OAuth and not with SharedKeys - and our STF tests
were inserting null into this field (which is fully valid from that
perspective). The code of the upgrader is already written to do "best
effort" extraction of tenantId and set a default empty string if needed.
It just not written assuming that even the authTokenEndpoint can be
null.

Change-Id: I9bb439ea8b1f6fe4cc1f0f8de08cbe33e5309ded
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36463
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Reviewed-by: Anthony Gatti <anthonyg@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit d2c1e4b74e8be2cfb5fa0d672138a6a976379dd3)
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36472