SDC-16315. Ability to configure refresh millis for JKS Credential store.
The default wait time before which the keystoer will be reloaded by JKS currently is harcoded to 10000 milliseconds, though this is a good default, in case of automation specifically for STF-1428 which will leverage JKs for auto credential creation to test multiple stages with credentials, configuring the reload time will help run tests correctly and with in less reload times.

This patch adds a credentialStore.jks.config.keystore.file.min.refresh.millis property so the minimum reload time for JKS keystore file can be configured.

Testing:
1. Added unit tests
2. Manually tested SDC with JKS with higher refresh time.

Change-Id: If404624e9b7726e0563057e06413faca33d841e0
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36556
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Sridhar Banoor <srid@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>