SDC-16320: FTP connector fails to close input stream
We will ignore the failure to close the input stream.
The servers we tested return an error code which causes an exception
to be thrown and caught here.

Normally we would need to issue an ABOR command, unfortunately
1) it's not exposed by the FTP client wrapper
2) anyway the current implementation in the library we use doesn't
send the command but closes a connection to the FTP server.

We do not expect any negative effect here, since in reality the data
connection is closed, the error the client throws is because the
server returns an error code for the previous command, that is RETR.
This is expected since we do not want to download the whole file,
we just wanted to see if we could read the content. This is the most
reliable way to test access permissions.

The worst case scenario: server losses a data connection and doesn't
finish (why wouldn't it?) the RETR command. In this case we wouldn't
see any exception here but the server would ignore any subsequent
control commands and the client may get stuck next time we try to
open a data connection. This is just a theoretical possibility which
wasn't yet seen in live. Since a workaround requires a lot of effort
and it's not clear if any of the customers uses incompatible FTP
servers, it was decided not to overcomplicate the solution but only
to log debug messages to help to identify if we hit the case. If that
happens we will get back to this issue to find another solutions.

Change-Id: Id9ce4be74f5f640ffcc503b744037d4c05388db2
Reviewed-on: https://review.streamsets.net/c/datacollector/+/36579
Reviewed-by: Hugo Prol <hugo@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>