SDC-16364. Delayed Queue can cause memory leak by holding references to closed record writers.
Delayed Queue can hold references to closed record writers. The HdfsTarget code uses DelayedQueue to hold writers to be closed after certain time, but when a RecordWriter is closed, we are not purging the items in the delayed queue, this can cause memory leak and consequently cause OOM. For ex: if u select the directoryInHeader configuration, the expiry time is Long.MAX_VALUE.

The RecordWriter also holds a reference to closed outputstream causing memory leak if the output stream implementation(ex: AbfsOutputStream) holds reference to byte array or a pool of byte arrays. This patch sets the output stream to null so it can be GCed as well as set expires to immediate so deleayed queue can release references to the writer.

Testing:
1. Before this patch saw a heap dump where there are as many DelayedRecordWriters with references to OutputStream having a reference to pool of byte arrays, after the patch, don't see those referenace, also memory does not slowly creep up after the patch.

Change-Id: I21aebfd4076f49020839a14c90347e663036e9f1
Reviewed-on: https://review.streamsets.net/c/datacollector/+/38084
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>