SDC-16388: Propagate DatabaseVendor in JDBC Lookup processor down to the shared libraries
While ago we've introduced concept of DatabaseVendor so that we can do
some database-specific mappings. I was working on standard tests for the
JDBC Lookup Processor and realized that the processor is not using the
vendor and thus some of the mappings are subject to old bugs such as the
one tracked in SDC-15071.

Change-Id: Ic0cc31b62bd6a390cd0a5957ef3d6dc9801b8eab
Reviewed-on: https://review.streamsets.net/c/datacollector/+/38147
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>