SDC-16357: all tests in stage.test_postgres_cdc are failing on upgrade path
The root issue is covered by SDC-16509 - basically our Yaml upgrader
does not respect types. The EL value() returns string which is then
stored in the configs, but this field is int and that is why the upgrade
tests are failing.

I've tried to load the pipeline in the UI and the thing is that
Javascript is not that type concerned so it happily casts the string to
int and then everything works. But it does require for someone to open
the pipeline in the UI which is not "good enough".

Changing the Yaml upgrader is a bit above my skill set and true to be
told, I don't want to touch it day before code freeze so I opted for
more a controversial approach - convert the PostgreSQL upgrader from
Yaml to Java where I can control types directly.

Change-Id: I9083f8b3195fd27d07c068449298955dc35bdb2b
Reviewed-on: https://review.streamsets.net/c/datacollector/+/38390
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>