SDC-16632: Univocity: Initial implementation
Initial implementation of the Univocity CSV parser. Exposed as an
explicit parser option with it's own set of configs (since the parsers
do differ in edge cases and the way some configs are interpreted).

It looks like many many changes, but true to be told, the main parser is
one file CsvUnivocityDelimitedParser and the test file. The rest is
wiring configuration for the UI.

I did some exploratory perf testing and performance wise for 4 columns,
the speed is about the same. The wider the record is, the more noticable
speed bump univocity gets.

Change-Id: I0a7eccc591a302e7251400c36e95369ec679c1ce
Reviewed-on: https://review.streamsets.net/c/datacollector/+/38963
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>