SDC-16628. Replace MySQL bin log connection with MySQL connection
We decided to implement a generic MySQL connection
to be used no only by the branded MySQL origin in Transformer
but also by the MySQL Binary log origin in Data Collector.

In this changset we replace host and port with one connectionString
configuration param. as mentioned above this affects the existing
MySQL Binary log origin and its YAML upgrader which is now a little
bit more complected, thus we added more unit tests for the upgrader.
Please also notice that the previous version wasn't release to
customers, thus we can reuse the same origin version, there is no need
to increase it.

Change-Id: If04c0bbb1519e2d7492b826864351b12236a57ac
Reviewed-on: https://review.streamsets.net/c/datacollector/+/38994
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>