SDC-16640. MySQL binary log tests are failing after addition of connection catalog support
We replaced the host and port params with the connection string param,
but we still need those individual parameters, thus we parse
the connection string param and extract the individual params.

Unfortunately the previous implementation parsed the connection string
incorrectly using the URI class. But URLs like
jdbc:mysql://MySQL_8.0.cluster:3306/default
are not parsed correctly by URI, the host will be null.
Instead of URI we can use URL which parses URLs as expected.
The only inconvenience with the URL class is that
a protocol handler should be installed. jdbc:mysql: most likely
doesn't exist at all. A solution is to replace jdbc:mysql: with file:
The file protocol handler is available by default.

Change-Id: I15de68948f46fd3a92e5497984b835ae8c7bb4af
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39054
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>