SDC-16651: Insert several thread dumps to support bundle
Got a good feedback from support folks that having by default several
minutes apart thread dumps in support bundle is good but - for several
use cases that is too far apart. Thus this patch further builds on top
of the last one and will by default insert three thread dumps to the
support bundle - taken during the bundle creation with 1 second wait in
between. Since  this way we wait only 2 seconds (three thread dumps
spaced 1 second), the delay is not that bad in terms of UX.

Change-Id: I433ee9e28090190e589e3f72243c5cdb1a6a5dfe
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39167
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>