SDC-16634: Univocity: Create upgraders in all components using the data parser capability
This patch adds explicit upgrader to each and every stage that uses the
config bean for data parser. Well almost, I have to deal with HDFS
Cluster mode separately. It's also not as large as you would guess as
lot of stages are using the services concept (a bit simpler).

Change-Id: I72d14ec08fee4b37148eb178461374b253d56d07
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39216
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>