SDC-16661: Univocity: Remove DataFormatParser class
It seems that this class is pretty much not used any more at all - it's
just referenced for one constant that doesn't even belong to this class.
So I've updated the references and dropped it.

Doing a bit of a digging, I believe that this class become unused as we
transitioned some of the stages to services concept.

Change-Id: Ib6ffb66fc0a8761e5279e71bc45cd5180235891d
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39219
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>