SDC-16672: Univocity: Switch default parser back to original one
Having more insights into how the parser works and talking to product
team, we've decided to switch the default back to the original parser.

The reason is that even though that Univocity is faster and thus better,
it doesn't come for free. First of all it will allocate memory based on
the max number of columns - I guess that is how it gets parts of it's
speed. But then and more importantly on the first parsing error it stops
- and there isn't really a good way to make it continue pass that point.
The original parser we have does have some limited capability to recover
(for example from long lines), that univocity does not.

Change-Id: If406e18492543ad2ec43fea439a083a65fb1a0f1
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39392
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>