SDC-16673: Create REST interface to generate list of stacktraces for support
Idea is to allow our support team to have a REST interface that they can
call to generate a set of stack traces with time delay in between. Of
course exposing such REST is running into the problem that if the SDC
itself is not healthy, this won't help. However there are scenarios
where access to shell is hard or where we need to observer certain
pipelines and SDC itself is responsive - thus the patch does make sense
in several use cases.

Change-Id: Ie54a4329c5221cf32b1a8320824072d6ce94f532
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39396
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>