SDC-16682. Revert "SDC-15864. Modify configuration value for record time live in writer"
This reverts commit 02b969ac200e79f5a079d2cfc5541cb8e0a93039.

Reason for revert: The SDC-15864 patch included an incorrect modification of the lifetime of the files created in the Local FS stage. This change introduced the value of the Idle Timeout property in the calculation of the record time, instead of fixing the behavior of Late Record Time Limit with respect to Idle Timeout as intended. After verifying that the behavior of Local FS is now incorrect and performing several configuration tests, the reason for the patch and therefore the ESC-978 was a misconfiguration and did not require any fix.

Change-Id: Ie816525fa572135f7a40bb5078e62e4567822ff0
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39424
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>