SDC-16680. Race condition when runner is removed
Before the runner was removed in a background thread. However there
could be a race condition where the runner is removed when pipeline is
about to start.
Here we are making sure the runner is removed as part of the state
transition hook so stopping a pipeline should remove the runner from
cache.

Change-Id: I728fa712be6915a85595f667e9ace4e988c11f45
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39734
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>