SDC-16769: Some UNIVOCITY tests are failing
CDH 6's version of Spark SQL depends on Univocity parser too and maven
happens to prefer to older version 2.7.x instead of picking up the 2.9.
that we need. Hence the exception with non-existing method.

Looking into univocity release notes:

https://github.com/uniVocity/univocity-parsers/releases

There is no backward incompatible change, so I decided to skip the
dependency from Spark and force maven to pick up 2.9 that we bring.

Change-Id: I51f7398590883e2c92968fc4007672ed552bafab
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39915
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>