SDC-16768. Better adjustment of LogMiner window limits in Oracle CDC
When CONTINUOUS_MINE is disabled and the dictionary source is
DICT_FROM_REDO_LOG, we need to manually register all the redo log files covering
from the dictionary beginning to the upper bound for the current mining
window. However, we do not need to expand the time window backwards to cover all
these redo log files – we can keep the same interval. This can help to speed up
the LogMiner session initialization.

Change-Id: I13a86dedb2f1a5f2e140175e9f3b937aff35620d
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39936
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Toni Martinez <toni@streamsets.com>