SDC-16779. Performance degradation in Oracle CDC
The regression is caused by SDC-15268. Prior to SDC-15268, Oracle CDC computed
the database time by converting the SDC's host time to the database
timezone. This could cause potential data loss (for Oracle 19c onwards) when the
database and SDC's host clocks are not synchronized. So the SDC-15268 fix was
introduced to directly query Oracle to get the database time. However, there is
a UI metric ("delay") that uses the database time – and that metric is updated
thousands of times per second (exactly one per each CDC record read). This is
what is causing the performance drop.

The solution would be to use the SDC's host time conversion to compute the
database time for the UI metric (to avoid the performance drop) while keeping
querying the Oracle database time for the mining process (to avoid data
loss). Regarding the second point, it will not cause performance degradation, as
the Oracle CDC mining process only queries the database time once per mining
window.

Change-Id: I2ac8d810ba6832d861e88f286a82e81224e0c6f8
Reviewed-on: https://review.streamsets.net/c/datacollector/+/39964
Reviewed-by: Toni Martinez <toni@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>