SDC-16837. Update the cluster config to support EMR Bootstrap Actions
Added Bootstrap Action configurations to the Transformer pipeline
configurations.

In DataCollector the same configurations are hidden,
but due to the UI limitations the Bootstrap tab will be always shown
but disabled.

Added Bootstrap Action configurations to the connection editor.

Pipeline and Connection upgraders were implemnted.

Added appropriate validations to the Connection and Pipeline editors

Change-Id: I7b0518d8d862227433bee3ffa2f723ac23a358de
Reviewed-on: https://review.streamsets.net/c/datacollector/+/40387
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Dimas Cabré i Chacón <dimas.cabre.chacon@streamsets.com>