SDC-16847. Oracle CDC raises JDBC_600 'missing redo log files error' when database and SDC are in different timezones
We used CURRENT_TIMESTAMP AT TIME ZONE DBTIMEZONE as the NEXT_TIME value
configured for the CURRENT online redo log in the LogMinerSession
queries. However, CURRENT_TIMESTAMP carries timezone info (it's an Oracle
TIMESTAMP WITH TIMEZONE value), while the rest of NEXT_TIME values don't (they
are Oracle DATE values).

LogMinerSession implementation expects DATE values expressed in local database
time, so that they can be directly converted from Timestamps to
LocalDateTimes. When doing so over a TIMESTAMP WITH TIMEZONE type, the result is
wrong -- the time ends up being expressed as a local time in the configured
JVM's timezone.

This patch avoids using CURRENT_TIMESTAMP (i.e. a TIMESTAMP WITH TIMEZONE value)
in favor of LOCALTIMESTAMP (i.e. a DATE value). Note that in order for this to
work, the DB session is configured with the DB timezone.

Change-Id: Id7b928a4deb44529b362321d6f112a91b1bc43b6
Reviewed-on: https://review.streamsets.net/c/datacollector/+/40392
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>