SDC-16866. Oracle CDC: LogMiner session can silently fail and provoke data loss on high throughput conditions
On high throughput load conditions and depending on specific redo log
settings (e.g. number and size of the online redo logs), Oracle can perform
several rotations of online redo logs in a short period of time. If an active
LogMiner session is processing one of the affected online logs, the session can
silenty fail – it won't retrieve more data and could finish with no distinction
from a normal "no more pending data" scenario.

In order to solve this situation, Oracle CDC needs to check the integrity of the
LogMiner session after finishing the mining process for the current window (this
involves comparing redo log metadata from before and after the LogMiner session
for all the online redo logs, if any, used in that mining window). And if the
integrity has potentially been compromised, Oracle CDC must replay the process
for the current mining window.

Change-Id: If6af8c351ed7a8f1f229f2e2054004ff789b7990
Reviewed-on: https://review.streamsets.net/c/datacollector/+/40524
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Toni Martinez <toni@streamsets.com>