SDC-16876. Oracle CDC: exclude redo logs not belonging to the current database incarnation
On databases with multiple incarnations, we should ensure that the redo logs
registered in LogMiner belong to the current incarnation. Otherwise, when
starting from a SCN or timestamp which is present in several incarnations,
Oracle CDC could load redo log files from the wrong incarnation, leading to
LogMiner start failures or wrong CDC data returned.

Change-Id: I140c99d8ad7af8b20a59f632289306fa496c3dac
Reviewed-on: https://review.streamsets.net/c/datacollector/+/40633
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Dimas Cabré i Chacón <dimas.cabre.chacon@streamsets.com>
Reviewed-on: https://review.streamsets.net/c/datacollector/+/40974
Reviewed-by: Hugo Prol <hugo@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>